Harden module flows and expand coverage
Add safer /back warmup/cooldown handling, public homes support, inventory writeback guards, and broader module tests.

+++ src/main/java/com/hanielcota/essentials/modules/chat/service/CooldownService.java
@@ -46,7 +46,7 @@ public final class CooldownService {
       return 0L;
     }
 
-    var cooldownMillis = (long) cooldownSeconds * 1_000L;
+    var cooldownMillis = cooldownSeconds * 1_000L;
     var now = System.currentTimeMillis();
     var elapsed = now - lastSend;
     var remaining = cooldownMillis - elapsed;
